+++ examples/tables.data.js
@@ -6,11 +6,46 @@ export const SIMPLE_TABLE = {
     { key: 'phone', content: 'Telefon' },
     { key: 'email', content: 'E-post'},
   ],
-  data: [
-    [ 'Ola Normann', 'Gateveien 2', '912 34 567', 'ola@normann.no' ],
-    [ 'Sivert Svenska', 'Gatuveio 7', '+46 123 456 789', 'sivert@svenska.se' ],
-    [ 'Daniel Dansk', 'Kartoffel Fjers', '+45 018 456 789', 'daniel@dansk.dk' ],
-    [ 'Anders Amerikaner', 'Road Street 52', '+1 555 555 5555', 'anders@amerikaner.biz' ],
+  dataObject: [
+    {
+      name: 'Ola Normann',
+      address: 'Gateveien 2',
+      age: 23,
+      phone: '912 34 567',
+      email: 'ola@normann.no',
+      username: 'olli22',
+    },
+    {
+      name: 'Sivert Svenska',
+      address: 'Gatuveio 7',
+      age: 45,
+      phone: '+46 123 456 789',
+      email: 'sivert@svenska.se',
+      username: 'zlatan_2000',
+    },
+    {
+      name: 'Daniel Dansk',
+      address: 'Kartoffel Fjers',
+      age: 67,
+      phone: '+45 018 456 789',
+      email: 'daniel@dansk.dk',
+      username: 'kartoffel88',
+    },
+    {
+      name: 'Anders Amerikaner',
+      address: 'Road Street 52',
+      age: 78,
+      phone: '+1 555 555 5555',
+      email: 'anders@amerikaner.biz',
+      username: 'andy-mandy',
+    },
+  ],
+};
+
+export const SIMPLE_TABLE_EXPANDABLE = {
+  headers: [
+    { key: 'name', content: 'Navn' },
+    { key: 'username', content: 'Brukernavn' }
   ],
   dataObject: [
     {
@@ -19,7 +54,7 @@ export const SIMPLE_TABLE = {
       age: 23,
       phone: '912 34 567',
       email: 'ola@normann.no',
-      username: 'olli22'
+      username: 'olli22',
     },
     {
       name: 'Sivert Svenska',
@@ -27,7 +62,7 @@ export const SIMPLE_TABLE = {
       age: 45,
       phone: '+46 123 456 789',
       email: 'sivert@svenska.se',
-      username: 'zlatan_2000'
+      username: 'zlatan_2000',
     },
     {
       name: 'Daniel Dansk',
@@ -35,7 +70,7 @@ export const SIMPLE_TABLE = {
       age: 67,
       phone: '+45 018 456 789',
       email: 'daniel@dansk.dk',
-      username: 'kartoffel88'
+      username: 'kartoffel88',
     },
     {
       name: 'Anders Amerikaner',
@@ -43,7 +78,7 @@ export const SIMPLE_TABLE = {
       age: 78,
       phone: '+1 555 555 5555',
       email: 'anders@amerikaner.biz',
-      username: 'andy-mandy'
+      username: 'andy-mandy',
     },
   ],
 };