+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 
+## v1.0.0
+
+* Responsive table
+* Sortable table
+* Expandable row
+* Update dependency ffe-tables to ^3.0.0
+
 ## v0.1.1
+
 * Updgrade eslint to use latest FFE configuration
 
 ## v0.1.0